# docs/api-types.ts
// Generated TypeScript types from HuleEdu API Gateway OpenAPI specification
// Do not edit this file manually - regenerate using scripts/generate_typescript_types.py

// ============================================================================
// API Response Types
// ============================================================================

export interface BatchClientStatus "available" | "processing"

export interface BatchPipelineRequest {
  batch_id?: string | any;
  requested_pipeline: PhaseName;
  is_retry?: boolean;
  retry_reason?: string | any
}

export interface BatchStatusResponse {
  status: BatchClientStatus;
  details: Record<string, any>
}

export interface HTTPValidationError {
  detail?: ValidationError[]
}

export interface PhaseName "spellcheck" | "ai_feedback" | "cj_assessment" | "nlp" | "student_matching"

export interface ValidationError {
  loc: string | number[];
  msg: string;
  type: string
}

// ============================================================================
// API Endpoint Types
// ============================================================================

export interface healthcheckhealthzgetResponse Record<string, string | Record<string, any>>

export interface proxyclassrequestsv1classespathpostResponse any

export interface proxyclassrequestsv1classespathpostResponse any

export interface proxyclassrequestsv1classespathpostResponse any

export interface proxyclassrequestsv1classespathpostResponse any

export interface testnoauthv1testnoauthgetResponse any

export interface testwithauthv1testwithauthgetResponse any

export interface getbatchstatusv1batchesbatchidstatusgetResponse BatchStatusResponse

export interface requestpipelineexecutionv1batchesbatchidpipelinespostRequest BatchPipelineRequest

// ============================================================================
// Common API Types
// ============================================================================

export interface ApiError {
  detail: string;
  status_code?: number;
  correlation_id?: string;
}

export interface ApiResponse<T> {
  data?: T;
  error?: ApiError;
  success: boolean;
}

// Authentication types
export interface JWTToken {
  access_token: string;
  token_type: string;
  expires_in?: number;
}

export interface AuthHeaders {
  Authorization: string;
  'X-Correlation-ID'?: string;
}

// WebSocket types
export interface WebSocketMessage {
  type: string;
  data: any;
  timestamp: string;
}

export interface TeacherNotification {
  teacher_id: string;
  notification_type: string;
  category: 'BATCH_PROCESSING' | 'CLASS_MANAGEMENT' | 'SYSTEM_ALERT' | 'USER_ACTION';
  priority: 'LOW' | 'STANDARD' | 'HIGH' | 'IMMEDIATE';
  payload: Record<string, any>;
  action_required: boolean;
  correlation_id: string;
  batch_id?: string;
  class_id?: string;
  timestamp: string;
}